requirements (1).txt
streamlit
python-dotenv
langchain-huggingface
langchain
huggingface-hub
